Finished the gaming page!
Made the wikipedia fetcher more generic so it works for any list of items (books or games).
Also now fetching the images from wikipedia and hosting them here. This will help if I need to add one off images, or want to replace any images myself.

Completed the layout of the gaming page